Fixed Inventory System Bugs
Fixed Inventory System Bug where items couldn't be transferred to and from the storage actor. The bug was created by the way items where added and removed from the inventory & was solved by implementing a new TransferItemTo function.

I also improved the comments in most of the classes making them more understandable for less experienced devs since this is a template & also generated a basic documentation pdf that still needs a lot of work.